refactor:util中collector改为point
