+++ src/main/java/cn/edu/sjtu/ist/ecssbackendedge/entity/domain/deviceController/DeviceController.java
@@ -1,0 +1,8 @@
+package cn.edu.sjtu.ist.ecssbackendedge.entity.domain.controller;
+
+/**
+ * @author dyanjun
+ * @date 2021/12/26 0:03
+ */
+public class deviceController {
+}
