+++ src/main/java/cn/edu/sjtu/ist/ecssbackendedge/entity/dto/request/DeviceControllerRequest.java
@@ -1,0 +1,8 @@
+package cn.edu.sjtu.ist.ecssbackendedge.entity.dto.request;
+
+/**
+ * @author dyanjun
+ * @date 2021/12/26 0:17
+ */
+public class DeviceControllerRequest {
+}
