+++ src/main/java/cn/edu/sjtu/ist/ecssbackendedge/entity/dto/response/DeviceControllerResponse.java
@@ -1,0 +1,8 @@
+package cn.edu.sjtu.ist.ecssbackendedge.entity.dto.response;
+
+/**
+ * @author dyanjun
+ * @date 2021/12/26 0:15
+ */
+public class DeviceControllerResponse {
+}
