+++ src/main/java/cn/edu/sjtu/ist/ecssbackendedge/entity/po/deviceController/DeviceControllerPO.java
@@ -1,0 +1,8 @@
+package cn.edu.sjtu.ist.ecssbackendedge.entity.po.deviceController;
+
+/**
+ * @author dyanjun
+ * @date 2021/12/26 0:18
+ */
+public class deviceControllerPO {
+}
