+++ src/main/java/cn/edu/sjtu/ist/ecssbackendedge/utils/convert/DeviceControllerUtil.java
@@ -1,0 +1,8 @@
+package cn.edu.sjtu.ist.ecssbackendedge.utils.convert;
+
+/**
+ * @author dyanjun
+ * @date 2021/12/26 0:19
+ */
+public class DeviceControllerUtil {
+}
