+++ src/main/java/cn/edu/sjtu/ist/ecssbackendedge/utils/point/zigbee/ZigBeeUtil.java
@@ -1,4 +1,4 @@
-package cn.edu.sjtu.ist.ecssbackendedge.utils.collect.zigbee;
+package cn.edu.sjtu.ist.ecssbackendedge.utils.point.zigbee;
 
 import cn.edu.sjtu.ist.ecssbackendedge.config.ZigBeeConfig;
